continue refactoring. now use_phonopy is much simplified
